Add Budget Link Value to DB

+++ Biblioteka/parametry_anteny.cs
@@ -12,7 +12,7 @@ namespace Biblioteka
         public string nazwa { get; set; }
         public double zysk_dBi { get; set; }
 
-        public double czestotliwosc_MHz { get; set; }
+        public int czestotliwosc_MHz { get; set; }
 
         /*public string parametry_anteny_opis
         {
